Finished but error (more in extended description)
I know that the error is I still have a letter print even if it doesn't fit because I check the square ahead of it as I go. I tried to make a helper that checked all the spaces ahead but only made it worst. Will fix later after asking some questions.